requirements.txt
aiobotocore==2.13.0
aiohttp==3.10.11
alembic==1.13.1
asyncpg==0.29.0
black==24.4.2
botocore==1.34.106
fastapi==0.111.0
httpx==0.27.0
pydantic==2.7.3
pydantic-settings==2.3.1
pytest==8.2.2
pytest-asyncio==0.23.7
python-dotenv==1.0.1
python-multipart==0.0.9
SQLAlchemy==2.0.30
uvicorn==0.30.1
